PERA-1540 Move BackupProtocolConstants into utils folder
