+++ common-sdk/src/main/kotlin/com/algorand/wallet/asb/domain/mapper/AsbBackupDataMapperImpl.kt
@@ -16,8 +16,8 @@ import com.algorand.wallet.asb.domain.model.AsbBackupAccount
 import com.algorand.wallet.asb.domain.model.AsbBackupData
 import com.algorand.wallet.asb.domain.model.BackupProtocolElement
 import com.algorand.wallet.asb.domain.model.BackupProtocolPayload
-import com.algorand.wallet.asb.domain.usecase.BackupProtocolConstants.SINGLE_ACCOUNT_TYPE_NAME
-import com.algorand.wallet.asb.domain.usecase.BackupProtocolConstants.WATCH_ACCOUNT_TYPE_NAME
+import com.algorand.wallet.asb.domain.utils.BackupProtocolConstants.SINGLE_ACCOUNT_TYPE_NAME
+import com.algorand.wallet.asb.domain.utils.BackupProtocolConstants.WATCH_ACCOUNT_TYPE_NAME
 import javax.inject.Inject
 
 internal class AsbBackupDataMapperImpl @Inject constructor() : AsbBackupDataMapper {
